remove redundent as

+++ src/main.rs
@@ -51,7 +51,7 @@ fn main() {
     hk.register_hotkey(modifiers::SHIFT, keys::CAPS_LOCK, pause)
         .unwrap();
     //115 is f4
-    hk.register_hotkey(alt_shift | modifiers::CONTROL, 115 as u32, || {
+    hk.register_hotkey(alt_shift | modifiers::CONTROL, 115_u32, || {
         std::process::exit(0x100)
     })
     .unwrap();
